# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Полное API для LocalVercel - локальной платформы для деплоя приложений\nПоддерживает проекты, деплои, логи, домены, переменные окружения,\nTelegram ботов, сборки, метрики, уведомления, Git интеграцию и шаблоны.\n\n## Аутентификация\nДля доступа к защищенным эндпоинтам используйте Bearer токен в заголовке:\n`Authorization: Bearer \u003cваш_токен\u003e`\n\n## WebSocket\nДля реального времени доступны WebSocket соединения:\n- `ws://localhost:8888/ws/deployments` - обновления деплоев\n- `ws://localhost:8888/ws/logs` - все логи\n- `ws://localhost:8888/ws/projects` - обновления проектов\n- `ws://localhost:8888/ws/logs/{deploymentId}` - логи конкретного деплоя",
        "title": "LocalVercel API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.localvercel.com/support",
            "email": "support@localvercel.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8888",
    "basePath": "/",
    "paths": {
...
        "/{resource}/me": {
            "get": {
                "description": "Возвращает ресурс текущего пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resources"
                ],
                "summary": "Получить ресурс",
                "responses": {
                    "200": {
                        "description": "Ресурс получен",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Обновляет ресурс текущего пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resources"
                ],
                "summary": "Обновить ресурс",
                "parameters": [
                    {
                        "description": "Данные для обновления",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ресурс обновлен",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "ok": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}